+++ CHANGES.md
@@ -4,6 +4,9 @@ See [compatibility docs](docs/compatibility.md) for details about backward compa
 
 ## CHANGE LOG ##
 
+* @dev
+   * Enforced HTTP code for error messages fits 400-499 range
+
 * v4.0.1 (2017-04-10)
    * TestingHelpers trait's `validateResponseStructure()` method is now public
    * [RB-64] Fixed Exception Handler generated HTTP code being out of allowed range in some cases